Rename & split file
